requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements-dev.txt pyproject.toml
#
attrs==22.2.0
    # via pytest
black==22.12.0
    # via livecode-archive (pyproject.toml)
cachetools==5.2.0
    # via google-auth
certifi==2022.12.7
    # via requests
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   black
    #   typer
contextlib2==21.6.0
    # via schema
detect-secrets==1.4.0
    # via livecode-archive (pyproject.toml)
distlib==0.3.6
    # via virtualenv
docopt==0.6.2
    # via internetarchive
exceptiongroup==1.1.0
    # via pytest
ffmpeg-python==0.2.0
    # via livecode-archive (pyproject.toml)
filelock==3.8.2
    # via virtualenv
future==0.18.2
    # via ffmpeg-python
google-api-core==2.11.0
    # via google-api-python-client
google-api-python-client==2.70.0
    # via livecode-archive (pyproject.toml)
google-auth==2.15.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==0.8.0
    # via livecode-archive (pyproject.toml)
googleapis-common-protos==1.57.0
    # via google-api-core
httplib2==0.21.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
identify==2.5.11
    # via pre-commit
idna==3.4
    # via requests
iniconfig==1.1.1
    # via pytest
internetarchive==3.0.2
    # via livecode-archive (pyproject.toml)
isort==5.11.4
    # via livecode-archive (pyproject.toml)
jsonpatch==1.32
    # via internetarchive
jsonpointer==2.3
    # via jsonpatch
libcst==0.4.9
    # via pycln
mypy-extensions==0.4.3
    # via
    #   black
    #   typing-inspect
nodeenv==1.7.0
    # via pre-commit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==22.0
    # via pytest
pathspec==0.9.0
    # via
    #   black
    #   pycln
platformdirs==2.6.0
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.20.0
    # via livecode-archive (pyproject.toml)
protobuf==4.21.12
    # via
    #   google-api-core
    #   googleapis-common-protos
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycln==2.1.2
    # via livecode-archive (pyproject.toml)
pyparsing==3.0.9
    # via httplib2
pytest==7.2.0
    # via livecode-archive (pyproject.toml)
pyyaml==6.0
    # via
    #   detect-secrets
    #   libcst
    #   pre-commit
    #   pycln
requests==2.28.1
    # via
    #   detect-secrets
    #   google-api-core
    #   internetarchive
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.9
    # via google-auth
schema==0.7.5
    # via internetarchive
six==1.16.0
    # via
    #   google-auth
    #   google-auth-httplib2
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   pytest
tomlkit==0.11.6
    # via pycln
tqdm==4.64.1
    # via
    #   internetarchive
    #   livecode-archive (pyproject.toml)
typer==0.7.0
    # via pycln
typing-extensions==4.4.0
    # via
    #   libcst
    #   typing-inspect
typing-inspect==0.8.0
    # via libcst
unidecode==1.3.6
    # via livecode-archive (pyproject.toml)
uritemplate==4.1.1
    # via google-api-python-client
urllib3==1.26.13
    # via
    #   internetarchive
    #   requests
virtualenv==20.17.1
    # via pre-commit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
